+++ arviz/__init__.py
@@ -9,8 +9,24 @@ from matplotlib.colors import LinearSegmentedColormap
 from matplotlib.pyplot import register_cmap, style
 
 
-# Configure logging before importing arviz internals
-_log = logging.getLogger("arviz")
+class Logger(logging.Logger):
+    """Override Logger to avoid repeated messages."""
+
+    def __init__(self, name, level=logging.NOTSET):
+        super().__init__(name=name, level=level)
+        self.cache = []
+
+    def _log(
+        self, level, msg, *args, **kwargs
+    ):  # pylint: disable=signature-differs, arguments-differ
+        msg_hash = hash(msg)
+        if msg_hash in self.cache:
+            return
+        self.cache.append(msg_hash)
+        super()._log(level, msg, *args, **kwargs)
+
+
+_log = Logger("arviz")
 
 
 from .data import *
@@ -306,4 +322,4 @@ _mpl_cm("gray_r", list(reversed(_linear_grey_10_95_c0)))
 
 
 # clean namespace
-del os, logging, register_cmap, LinearSegmentedColormap
+del os, logging, register_cmap, LinearSegmentedColormap, Logger