+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Fix `pair_plot` for mixed discrete and continuous variables ([1434](https://github.com/arviz-devs/arviz/pull/1434))
 * Fix in-sample deviance in `plot_compare` ([1435](https://github.com/arviz-devs/arviz/pull/1435))
 * Fix computation of weights in compare ([1438](https://github.com/arviz-devs/arviz/pull/1438))
+* Avoid repeated warning in summary ([1442](https://github.com/arviz-devs/arviz/pull/1442)) 
 
 ### Deprecation